cleanup usings and switched content of files to match the file names
