+++ Core.DomainModel/ItSystemUsage/Read/ItSystemUsageOverviewItContractReadModel.cs
@@ -1,17 +1,14 @@
 ﻿using System;
-using Core.DomainModel.Shared;
 
 namespace Core.DomainModel.ItSystemUsage.Read
 {
-    public class ItSystemUsageOverviewDataProcessingRegistrationReadModel
+    public class ItSystemUsageOverviewItContractReadModel
     {
         public int Id { get; set; }
 
-        public Guid DataProcessingRegistrationUuid { get; set; }
-        public int DataProcessingRegistrationId { get; set; }
-        public string DataProcessingRegistrationName { get; set; }
-
-        public YesNoIrrelevantOption? IsAgreementConcluded { get; set; }
+        public Guid ItContractUuid { get; set; }
+        public int ItContractId { get; set; }
+        public string ItContractName { get; set; }
 
         public int ParentId { get; set; }
         public virtual ItSystemUsageOverviewReadModel Parent { get; set; }
