+++ Core.DomainServices/Repositories/SystemUsage/ItSystemUsageRepository.cs
@@ -2,7 +2,6 @@
 using System.Linq;
 using Core.DomainModel.ItSystemUsage;
 using Core.DomainServices.Extensions;
-using NotImplementedException = System.NotImplementedException;
 
 namespace Core.DomainServices.Repositories.SystemUsage
 {
